--- commands/automatic_move.py
@@ -1,1 +1,0 @@
-ss
